LPS-70711 Ensure all configurations are deleted on tearDown
Refactor common code into base class
