+++ adaptive-media-image-impl-test/src/testIntegration/java/com/liferay/adaptive/media/image/internal/test/ImageAdaptiveMediaConfigurationTest.java
@@ -17,18 +17,11 @@ package com.liferay.adaptive.media.image.internal.test;
 import com.liferay.adaptive.media.image.configuration.ImageAdaptiveMediaConfigurationEntry;
 import com.liferay.adaptive.media.image.configuration.ImageAdaptiveMediaConfigurationHelper;
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.registry.Filter;
-import com.liferay.registry.Registry;
-import com.liferay.registry.RegistryUtil;
-import com.liferay.registry.ServiceTracker;
-
-import java.io.IOException;
 
 import java.util.Collection;
 import java.util.HashMap;
@@ -36,11 +29,7 @@ import java.util.Iterator;
 import java.util.Map;
 import java.util.Optional;
 
-import org.junit.After;
-import org.junit.AfterClass;
 import org.junit.Assert;
-import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
@@ -51,7 +40,8 @@ import org.junit.runner.RunWith;
  */
 @RunWith(Arquillian.class)
 @Sync
-public class ImageAdaptiveMediaConfigurationTest {
+public class ImageAdaptiveMediaConfigurationTest
+	extends ImageAdaptiveMediaConfigurationBaseTest {
 
 	@ClassRule
 	@Rule
@@ -60,40 +50,12 @@ public class ImageAdaptiveMediaConfigurationTest {
 			new LiferayIntegrationTestRule(),
 			SynchronousDestinationTestRule.INSTANCE);
 
-	@BeforeClass
-	public static void setUpClass() {
-		Registry registry = RegistryUtil.getRegistry();
-
-		Filter filter = registry.getFilter(
-			"(objectClass=" +
-				ImageAdaptiveMediaConfigurationHelper.class.getName() + ")");
-
-		_serviceTracker = registry.trackServices(filter);
-
-		_serviceTracker.open();
-	}
-
-	@AfterClass
-	public static void tearDownClass() {
-		_serviceTracker.close();
-	}
-
-	@Before
-	public void setUp() throws Exception {
-		_deleteAllConfigurationEntries();
-	}
-
-	@After
-	public void tearDown() throws Exception {
-		_deleteAllConfigurationEntries();
-	}
-
 	@Test
 	public void testAddConfigurationEntryWithExistingDisabledConfiguration()
 		throws Exception {
 
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -137,7 +99,7 @@ public class ImageAdaptiveMediaConfigurationTest {
 	@Test
 	public void testEmptyConfiguration() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Iterable<ImageAdaptiveMediaConfigurationEntry> configurationEntries =
 			configurationHelper.getImageAdaptiveMediaConfigurationEntries(
@@ -152,7 +114,7 @@ public class ImageAdaptiveMediaConfigurationTest {
 	@Test
 	public void testExistantConfigurationEntry() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -175,7 +137,7 @@ public class ImageAdaptiveMediaConfigurationTest {
 		throws Exception {
 
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -218,7 +180,7 @@ public class ImageAdaptiveMediaConfigurationTest {
 		throws Exception {
 
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -264,7 +226,7 @@ public class ImageAdaptiveMediaConfigurationTest {
 		throws Exception {
 
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -288,7 +250,7 @@ public class ImageAdaptiveMediaConfigurationTest {
 	@Test
 	public void testNonEmptyConfiguration() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -308,7 +270,7 @@ public class ImageAdaptiveMediaConfigurationTest {
 	@Test
 	public void testNonExistantConfigurationEntry() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -325,27 +287,5 @@ public class ImageAdaptiveMediaConfigurationTest {
 
 		Assert.assertFalse(configurationEntryOptional.isPresent());
 	}
-
-	private void _deleteAllConfigurationEntries()
-		throws IOException, PortalException {
-
-		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
-
-		Collection<ImageAdaptiveMediaConfigurationEntry> configurationEntries =
-			configurationHelper.getImageAdaptiveMediaConfigurationEntries(
-				TestPropsValues.getCompanyId(), configurationEntry -> true);
-
-		for (ImageAdaptiveMediaConfigurationEntry configurationEntry :
-				configurationEntries) {
-
-			configurationHelper.forceDeleteImageAdaptiveMediaConfigurationEntry(
-				TestPropsValues.getCompanyId(), configurationEntry.getUUID());
-		}
-	}
-
-	private static ServiceTracker
-		<ImageAdaptiveMediaConfigurationHelper,
-			ImageAdaptiveMediaConfigurationHelper> _serviceTracker;
 
 }
