+++ adaptive-media-image-impl-test/src/testIntegration/java/com/liferay/adaptive/media/image/internal/test/ImageAdaptiveMediaDeleteConfigurationTest.java
@@ -23,20 +23,12 @@ import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.registry.Filter;
-import com.liferay.registry.Registry;
-import com.liferay.registry.RegistryUtil;
-import com.liferay.registry.ServiceTracker;
 
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 
-import org.junit.AfterClass;
 import org.junit.Assert;
-import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
@@ -47,7 +39,8 @@ import org.junit.runner.RunWith;
  */
 @RunWith(Arquillian.class)
 @Sync
-public class ImageAdaptiveMediaDeleteConfigurationTest {
+public class ImageAdaptiveMediaDeleteConfigurationTest
+	extends ImageAdaptiveMediaConfigurationBaseTest {
 
 	@ClassRule
 	@Rule
@@ -56,45 +49,10 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 			new LiferayIntegrationTestRule(),
 			SynchronousDestinationTestRule.INSTANCE);
 
-	@BeforeClass
-	public static void setUpClass() {
-		Registry registry = RegistryUtil.getRegistry();
-
-		Filter filter = registry.getFilter(
-			"(objectClass=" +
-				ImageAdaptiveMediaConfigurationHelper.class.getName() + ")");
-
-		_serviceTracker = registry.trackServices(filter);
-
-		_serviceTracker.open();
-	}
-
-	@AfterClass
-	public static void tearDownClass() {
-		_serviceTracker.close();
-	}
-
-	@Before
-	public void setUp() throws Exception {
-		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
-
-		Collection<ImageAdaptiveMediaConfigurationEntry> configurationEntries =
-			configurationHelper.getImageAdaptiveMediaConfigurationEntries(
-				TestPropsValues.getCompanyId(), configurationEntry -> true);
-
-		for (ImageAdaptiveMediaConfigurationEntry configurationEntry :
-				configurationEntries) {
-
-			configurationHelper.forceDeleteImageAdaptiveMediaConfigurationEntry(
-				TestPropsValues.getCompanyId(), configurationEntry.getUUID());
-		}
-	}
-
 	@Test
 	public void testDeleteAllConfigurationEntries() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -117,14 +75,14 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "1");
 
-		_assertEnabled(firstConfigurationEntryOptional);
+		assertEnabled(firstConfigurationEntryOptional);
 
 		Optional<ImageAdaptiveMediaConfigurationEntry>
 			secondConfigurationEntryOptional =
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "2");
 
-		_assertEnabled(secondConfigurationEntryOptional);
+		assertEnabled(secondConfigurationEntryOptional);
 
 		configurationHelper.disableImageAdaptiveMediaConfigurationEntry(
 			TestPropsValues.getCompanyId(), "1");
@@ -156,7 +114,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 		throws Exception {
 
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -188,7 +146,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "1");
 
-		_assertDisabled(configurationEntryOptional);
+		assertDisabled(configurationEntryOptional);
 
 		configurationEntryOptional =
 			configurationHelper.getImageAdaptiveMediaConfigurationEntry(
@@ -200,7 +158,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 	@Test
 	public void testDeleteDeletedConfigurationEntry() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -215,7 +173,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "1");
 
-		_assertEnabled(configurationEntryOptional);
+		assertEnabled(configurationEntryOptional);
 
 		configurationHelper.disableImageAdaptiveMediaConfigurationEntry(
 			TestPropsValues.getCompanyId(), "1");
@@ -245,7 +203,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 	)
 	public void testDeleteEnabledConfigurationEntry() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -260,7 +218,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "1");
 
-		_assertEnabled(configurationEntryOptional);
+		assertEnabled(configurationEntryOptional);
 
 		configurationHelper.deleteImageAdaptiveMediaConfigurationEntry(
 			TestPropsValues.getCompanyId(), "1");
@@ -269,7 +227,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 	@Test
 	public void testDeleteFirstConfigurationEntry() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -292,14 +250,14 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "1");
 
-		_assertEnabled(firstConfigurationEntryOptional);
+		assertEnabled(firstConfigurationEntryOptional);
 
 		Optional<ImageAdaptiveMediaConfigurationEntry>
 			secondConfigurationEntryOptional =
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "2");
 
-		_assertEnabled(secondConfigurationEntryOptional);
+		assertEnabled(secondConfigurationEntryOptional);
 
 		configurationHelper.disableImageAdaptiveMediaConfigurationEntry(
 			TestPropsValues.getCompanyId(), "1");
@@ -317,13 +275,13 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 			configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 				TestPropsValues.getCompanyId(), "2");
 
-		_assertEnabled(secondConfigurationEntryOptional);
+		assertEnabled(secondConfigurationEntryOptional);
 	}
 
 	@Test
 	public void testDeleteSecondConfigurationEntry() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -346,14 +304,14 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "1");
 
-		_assertEnabled(firstConfigurationEntryOptional);
+		assertEnabled(firstConfigurationEntryOptional);
 
 		Optional<ImageAdaptiveMediaConfigurationEntry>
 			secondConfigurationEntryOptional =
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "2");
 
-		_assertEnabled(secondConfigurationEntryOptional);
+		assertEnabled(secondConfigurationEntryOptional);
 
 		configurationHelper.disableImageAdaptiveMediaConfigurationEntry(
 			TestPropsValues.getCompanyId(), "2");
@@ -365,7 +323,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 			configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 				TestPropsValues.getCompanyId(), "1");
 
-		_assertEnabled(firstConfigurationEntryOptional);
+		assertEnabled(firstConfigurationEntryOptional);
 
 		secondConfigurationEntryOptional =
 			configurationHelper.getImageAdaptiveMediaConfigurationEntry(
@@ -377,7 +335,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 	@Test
 	public void testDeleteUniqueConfigurationEntry() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -392,7 +350,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "1");
 
-		_assertEnabled(configurationEntryOptional);
+		assertEnabled(configurationEntryOptional);
 
 		configurationHelper.disableImageAdaptiveMediaConfigurationEntry(
 			TestPropsValues.getCompanyId(), "1");
@@ -410,7 +368,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 	@Test
 	public void testForceDeleteAllConfigurationEntries() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -433,14 +391,14 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "1");
 
-		_assertEnabled(firstConfigurationEntryOptional);
+		assertEnabled(firstConfigurationEntryOptional);
 
 		Optional<ImageAdaptiveMediaConfigurationEntry>
 			secondConfigurationEntryOptional =
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "2");
 
-		_assertEnabled(secondConfigurationEntryOptional);
+		assertEnabled(secondConfigurationEntryOptional);
 
 		configurationHelper.forceDeleteImageAdaptiveMediaConfigurationEntry(
 			TestPropsValues.getCompanyId(), "1");
@@ -467,7 +425,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 		throws Exception {
 
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -496,7 +454,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "1");
 
-		_assertDisabled(configurationEntryOptional);
+		assertDisabled(configurationEntryOptional);
 
 		configurationEntryOptional =
 			configurationHelper.getImageAdaptiveMediaConfigurationEntry(
@@ -508,7 +466,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 	@Test
 	public void testForceDeleteDeletedConfigurationEntry() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -523,7 +481,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "1");
 
-		_assertEnabled(configurationEntryOptional);
+		assertEnabled(configurationEntryOptional);
 
 		configurationHelper.forceDeleteImageAdaptiveMediaConfigurationEntry(
 			TestPropsValues.getCompanyId(), "1");
@@ -547,7 +505,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 	@Test
 	public void testForceDeleteEnabledConfigurationEntry() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -562,7 +520,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "1");
 
-		_assertEnabled(configurationEntryOptional);
+		assertEnabled(configurationEntryOptional);
 
 		configurationHelper.forceDeleteImageAdaptiveMediaConfigurationEntry(
 			TestPropsValues.getCompanyId(), "1");
@@ -577,7 +535,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 	@Test
 	public void testForceDeleteFirstConfigurationEntry() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -600,14 +558,14 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "1");
 
-		_assertEnabled(firstConfigurationEntryOptional);
+		assertEnabled(firstConfigurationEntryOptional);
 
 		Optional<ImageAdaptiveMediaConfigurationEntry>
 			secondConfigurationEntryOptional =
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "2");
 
-		_assertEnabled(secondConfigurationEntryOptional);
+		assertEnabled(secondConfigurationEntryOptional);
 
 		configurationHelper.forceDeleteImageAdaptiveMediaConfigurationEntry(
 			TestPropsValues.getCompanyId(), "1");
@@ -622,13 +580,13 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 			configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 				TestPropsValues.getCompanyId(), "2");
 
-		_assertEnabled(secondConfigurationEntryOptional);
+		assertEnabled(secondConfigurationEntryOptional);
 	}
 
 	@Test
 	public void testForceDeleteSecondConfigurationEntry() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -651,14 +609,14 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "1");
 
-		_assertEnabled(firstConfigurationEntryOptional);
+		assertEnabled(firstConfigurationEntryOptional);
 
 		Optional<ImageAdaptiveMediaConfigurationEntry>
 			secondConfigurationEntryOptional =
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "2");
 
-		_assertEnabled(secondConfigurationEntryOptional);
+		assertEnabled(secondConfigurationEntryOptional);
 
 		configurationHelper.forceDeleteImageAdaptiveMediaConfigurationEntry(
 			TestPropsValues.getCompanyId(), "2");
@@ -667,7 +625,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 			configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 				TestPropsValues.getCompanyId(), "1");
 
-		_assertEnabled(firstConfigurationEntryOptional);
+		assertEnabled(firstConfigurationEntryOptional);
 
 		secondConfigurationEntryOptional =
 			configurationHelper.getImageAdaptiveMediaConfigurationEntry(
@@ -679,7 +637,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 	@Test
 	public void testForceDeleteUniqueConfigurationEntry() throws Exception {
 		ImageAdaptiveMediaConfigurationHelper configurationHelper =
-			_serviceTracker.getService();
+			serviceTracker.getService();
 
 		Map<String, String> properties = new HashMap<>();
 
@@ -694,7 +652,7 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 				configurationHelper.getImageAdaptiveMediaConfigurationEntry(
 					TestPropsValues.getCompanyId(), "1");
 
-		_assertEnabled(configurationEntryOptional);
+		assertEnabled(configurationEntryOptional);
 
 		configurationHelper.forceDeleteImageAdaptiveMediaConfigurationEntry(
 			TestPropsValues.getCompanyId(), "1");
@@ -705,33 +663,5 @@ public class ImageAdaptiveMediaDeleteConfigurationTest {
 
 		Assert.assertFalse(configurationEntryOptional.isPresent());
 	}
-
-	private void _assertDisabled(
-		Optional<ImageAdaptiveMediaConfigurationEntry>
-			configurationEntryOptional) {
-
-		Assert.assertTrue(configurationEntryOptional.isPresent());
-
-		ImageAdaptiveMediaConfigurationEntry configurationEntry =
-			configurationEntryOptional.get();
-
-		Assert.assertFalse(configurationEntry.isEnabled());
-	}
-
-	private void _assertEnabled(
-		Optional<ImageAdaptiveMediaConfigurationEntry>
-			configurationEntryOptional) {
-
-		Assert.assertTrue(configurationEntryOptional.isPresent());
-
-		ImageAdaptiveMediaConfigurationEntry configurationEntry =
-			configurationEntryOptional.get();
-
-		Assert.assertTrue(configurationEntry.isEnabled());
-	}
-
-	private static ServiceTracker
-		<ImageAdaptiveMediaConfigurationHelper,
-			ImageAdaptiveMediaConfigurationHelper> _serviceTracker;
 
 }
